Client/requirements.txt
llama-cpp-python
pyautogui
mss
pillow
python-xlib
